Add Title bar
Add the title bar, this adds all of the styling necessary to start the the title bar later. It is not the final location of the bar; however, the sizing should be correct.

Additionally ci.py is moved to the top of the repo and a unit test is added for get_title_menu_options.